+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/chat/components/AvatarView.kt
@@ -4,7 +4,6 @@
 package com.azure.android.communication.ui.chat.presentation.ui.chat.components
 
 import androidx.annotation.DrawableRes
-import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
@@ -12,7 +11,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.viewinterop.AndroidView
 import com.microsoft.fluentui.persona.AvatarSize
 import com.microsoft.fluentui.persona.AvatarView
-
 
 @Composable
 internal fun AvatarView(name: String?, color: Color? = null, avatarSize: AvatarSize = AvatarSize.LARGE, @DrawableRes image: Int = -1, isGrouped: Boolean = false) {
@@ -31,7 +29,5 @@ internal fun AvatarView(name: String?, color: Color? = null, avatarSize: AvatarS
 @Preview
 @Composable
 internal fun PreviewAvatarView() {
-    Column {
-        AvatarView(name = "John Doe")
-    }
+    AvatarView(name = "John Doe")
 }
