+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/container/ChatView.kt
@@ -26,21 +26,21 @@ internal class ChatView(context: Context, private val instanceId: Int) : FrameLa
     override fun onAttachedToWindow() {
         super.onAttachedToWindow()
         ReduxViewModelGenerator(
-                builder = { store ->
-                    buildChatScreenViewModel(
-                            store = store,
-                            repository = locator.locate()
-                    )
-                },
-                onChanged = {
-                    composeView.setContent {
-                        ChatCompositeTheme {
-                            ChatScreen(viewModel = it)
-                        }
+            builder = { store ->
+                buildChatScreenViewModel(
+                    store = store,
+                    repository = locator.locate()
+                )
+            },
+            onChanged = {
+                composeView.setContent {
+                    ChatCompositeTheme {
+                        ChatScreen(viewModel = it)
                     }
-                },
-                coroutineScope = findViewTreeLifecycleOwner()!!.lifecycleScope,
-                store = locator.locate()
+                }
+            },
+            coroutineScope = findViewTreeLifecycleOwner()!!.lifecycleScope,
+            store = locator.locate()
         )
     }
 }
