+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/viewmodel/MessageViewModel.kt
@@ -53,4 +53,3 @@ private class InfoModelToViewModelAdapter(private val messages: List<MessageInfo
         TODO("Not Implemented, probably not needed")
     }
 }
-
